# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_anzhiyu-c_anheyu-app_pkg_service_theme.ThemeInstallRequest:
    properties:
      download_url:
        type: string
      market_id:
        type: integer
      theme_name:
        type: string
      version:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  theme.MarketTheme:
    properties:
      author:
        type: string
      createdAt:
        type: string
      demoUrl:
        type: string
      description:
        type: string
      downloadCount:
        type: integer
      downloadUrl:
        type: string
      id:
        type: integer
      instructionUrl:
        type: string
      isActive:
        type: boolean
      isOfficial:
        type: boolean
      name:
        type: string
      previewUrl:
        type: string
      price:
        type: integer
      rating:
        type: number
      repoUrl:
        type: string
      tags:
        items:
          type: string
        type: array
      themeType:
        type: string
      updatedAt:
        type: string
      version:
        type: string
    type: object
  theme.StaticModeResponse:
    properties:
      is_active:
        type: boolean
    type: object
  theme.SwitchThemeRequest:
    properties:
      theme_name:
        type: string
    required:
    - theme_name
    type: object
  theme.ThemeInfo:
    properties:
      author:
        type: string
      createdAt:
        type: string
      demoUrl:
        type: string
      description:
        type: string
      downloadCount:
        type: integer
      downloadUrl:
        type: string
      id:
        type: integer
      install_time:
        description: 安装时间
        type: string
      installed_version:
        description: 已安装版本
        type: string
      instructionUrl:
        type: string
      is_current:
        description: 本地状态字段
        type: boolean
      is_installed:
        description: 是否已安装（对于已安装主题列表始终为true）
        type: boolean
      isActive:
        type: boolean
      isOfficial:
        type: boolean
      name:
        type: string
      previewUrl:
        type: string
      price:
        type: integer
      rating:
        type: number
      repoUrl:
        type: string
      tags:
        items:
          type: string
        type: array
      themeType:
        type: string
      updatedAt:
        type: string
      user_config:
        additionalProperties: true
        description: 用户配置
        type: object
      version:
        type: string
    type: object
  theme.ThemeMarketListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/theme.MarketTheme'
        type: array
      total:
        type: integer
    type: object
  theme.ThemeUploadResponse:
    properties:
      installed:
        type: boolean
      message:
        type: string
      theme_info: {}
      theme_name:
        type: string
    type: object
  theme.UninstallThemeRequest:
    properties:
      theme_name:
        type: string
    required:
    - theme_name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anheyu.com
    name: 安知鱼
    url: https://github.com/anzhiyu-c/anheyu-app
  description: Anheyu App 应用接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Anheyu App API
  version: "1.0"
paths:
...
  /public/theme/market:
    get:
      description: 获取主题商城中的所有可用主题
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.ThemeMarketListResponse'
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取主题商城列表
      tags:
      - 主题商城
...
  /theme/current:
    get:
      description: 获取用户当前使用的主题信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.ThemeInfo'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前主题
      tags:
      - 主题管理
  /theme/fix-status:
    post:
      description: 修复用户主题的当前状态数据一致性，解决多个主题同时标记为当前使用的问题
      produces:
      - application/json
      responses:
        "200":
          description: 修复成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 修复失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修复主题状态
      tags:
      - 主题管理
  /theme/install:
    post:
      consumes:
      - application/json
      description: 从指定URL下载并安装主题（主题名必须以theme-开头）
      parameters:
      - description: 主题安装请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_anzhiyu-c_anheyu-app_pkg_service_theme.ThemeInstallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 安装成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 安装失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 安装主题
      tags:
      - 主题管理
  /theme/installed:
    get:
      description: 获取用户已安装的所有主题
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/theme.ThemeInfo'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取已安装主题列表
      tags:
      - 主题管理
  /theme/official:
    post:
      description: 切换到官方内嵌主题
      produces:
      - application/json
      responses:
        "200":
          description: 切换成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 切换失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 切换到官方主题
      tags:
      - 主题管理
  /theme/static-mode:
    get:
      description: 检查当前是否处于静态主题模式（是否存在static目录）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.StaticModeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 检查静态模式
      tags:
      - 主题管理
  /theme/switch:
    post:
      consumes:
      - application/json
      description: 切换到指定的已安装主题或官方主题
      parameters:
      - description: 切换主题请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/theme.SwitchThemeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 切换成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 切换失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 切换主题
      tags:
      - 主题管理
  /theme/uninstall:
    post:
      consumes:
      - application/json
      description: 卸载指定的主题（不能卸载当前使用的主题）
      parameters:
      - description: 卸载主题请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/theme.UninstallThemeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 卸载成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 卸载失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 卸载主题
      tags:
      - 主题管理
  /theme/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传主题压缩包文件（ZIP格式，最大50MB），系统会自动解析theme.json并安装主题
      parameters:
      - description: 主题压缩包文件
        in: formData
        name: file
        required: true
        type: file
      - description: 是否强制更新
        in: formData
        name: force_update
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/theme.ThemeUploadResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 上传失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 上传主题压缩包
      tags:
      - 主题管理
  /theme/validate:
    post:
      consumes:
      - multipart/form-data
      description: 验证主题压缩包的格式和内容是否符合规范
      parameters:
      - description: 主题压缩包文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 验证成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 验证失败
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 验证主题压缩包
      tags:
      - 主题管理
...
securityDefinitions:
  BearerAuth:
    description: '在请求头中添加 Bearer Token，格式为: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"